修改几个编译警告。

+++ SOUI/src/core/shostwnd.cpp
@@ -1449,7 +1449,7 @@ LRESULT SHostWnd::OnScriptTimer( UINT uMsg,WPARAM wParam,LPARAM lParam )
     if(m_pScriptModule)
     {
         EventTimer evt(this);
-		evt.uID = wParam;
+		evt.uID = (UINT)wParam;
         m_pScriptModule->executeScriptedEventHandler((LPCSTR)lParam,&evt);
     }
     return 0;
